safelinux-system-cfg: devicetree: Add support for main-dts download
There is a need to download the main dts file, compile them with
overlay support, and apply the external/out of tree DTBO files to
the same. Hence, add the support for following:
1. Download the mainline dts code
2.  Compile the same with overlay option enabled.
3. Merge the dtb files to corresponding matching main dtb files.
matching criteria depends upon the properties 'compatible' and
'model' under root node, for dtbo and dtb file.

Change-Id: I71f240170b860544d3b4738eda5fa42cd23bad46
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>