hack: devicetree: sa7255p: keep the NSP's corner votes intact
NSP's corner votes from bootloader are removed as part of sync_state
and causing clock enablement failures from trusted firmware when PIL
loading is done for the NSP's from userspace.
So added a dummy DT node to keep the votes intact as a hack until scmi
migration is done for the sa7255p platform and uscmi is available for
the user mode PIL drivers to vote for these corners explicitly.

Change-Id: I14b00aedca5ea4d00606e411a4b82a1773a6f834
Signed-off-by: Shazad Hussain <quic_shazhuss@quicinc.com>