devicetree: sa8255p_overlay: remove the ipcc node
Remove the ipcc node in sa8255p_overlay.dtsi because it was
already defined in the kernel devicetree, meanwhile rename
ipcc to ipcc0 in sa8775p-vfio.dtsi.

To be compatible with sa8775 variant which is ARK based, create
sa8775p-vfio_rh.dtsi which is only be used for ARK kernel.

Change-Id: I1e58b11a0bd3e7337afe14c52880e9a06409c315
Signed-off-by: Haixu Cui <quic_haixcui@quicinc.com>