devicetree: move cma memory-region from shmbridge to scm_user_intf_sec
To let memory regions be accessed by TrustZone we need to register them
in TZ. The registeration was done using qtee_shmbridge driver, but we
migrate to do it using qcom_tzmem driver.

The CMA heap was registered in the qtee_shmbridge driver itself, and the
driver got the cma memory-region from a DeviceTree entry.
This entry is no longer needed in qtee_shmbridge node since it's about
to be deprecated.

qcom_tzmem doesn't register the cma heap as part of the driver, so we do
the registration in scm_user_intf_sec driver. This driver needs the
memory-region of the cma, similar to what was done for qtee_shmbridge
node.

Change-Id: I1cf07b497c3b937689f366b9da01eb45a3325a53
Signed-off-by: Yiftach Ariel <quic_yariel@quicinc.com>