arm64: dts: qcom: sa7255p: sa8775p: Add sail mailbox channel for power related use case
Add support for the MCU shutdown feature when the Vehicle Interface Processor (VIP) initiates the shutdown command. The shutdown command is acceptable once both the apps processor and the safety island OEM Mailbox communication reach the end to end ready state.
Additionally, power channel IPCC interrupts are added to handle scenarios where the main domain needs to communicate with the safety island. The IPCC interrupt associated with the power mailbox channel is added to the sail mailbox.

Change-Id: Ife13ae0dfd29b5f1707730b3852214a0778bdd48
Signed-off-by: Priyadarshini Kulkarni <quic_priykulk@quicinc.com>