vfio:refine restorecon judgement condition
As we run vfio script in early-ramdisk,
in this early stage the env is not ready.

selinuxenabled is a binary and can be used directly.
Refine the judgement condition to avoid selinux
avc deny because of wrong label.

Change-Id: I35eb9c46a5309e04760f02aaa979b37cd5e5bf65
Signed-off-by: ziqic <ziqic@qti.qualcomm.com>
(cherry picked from commit aaccc04f72e49daca28e05a4513d3b6700f20f96)