devicetree: register tzcom heap in shmbridge
Sharing dma-buffers with TZ is currently done by allocating the
dma-buffers from the linux,cma heap.
For safety concerns, to support TZ-FFI, we need a dedicated heap for
sharing dma-buffers with TZ.
This dedicated heap is called qcom,tzcom heap.

Sharing buffers with TZ requires them to be registered in shmbridge.
In this commit we replace the registration of the linux,cma heap in
shmbridge with the registration of the qcom,tzcom heap.
This means that buffer allocated from linux,cma heap can no longer be
shared with TZ. Instead, the buffers need to be allocated from
qcom,tzcom heap.

Change-Id: I3a87c063ccd6cef918701f7d101b7ce1030b153b
Signed-off-by: Yiftach Ariel <yariel@qti.qualcomm.com>