devicetree: sa8650p: Add overlay dt to configure non_ecc_heap_mem
Since target devices can have varying DDR configurations, the
memory allocation for the non_ecc_heap changes based on the DDR size
available on the target. To accommodate this, we create overlay
device trees for this.

ABL selects the appropriate device tree (DT) based on the
subtype value in the board-id. Bits 8-10 in the board-id
subtype field are reserved for DDR type. This subtype value
is updated during the overlay process.

Change-Id: I24d3f9ccb0a515a30c7f82bede261f294ab4168b
Signed-off-by: Wei Liu <wliu8@qti.qualcomm.com>